move chat stuff to new folder and format
