+++ gamedaybot/__init__.py
@@ -1,0 +1,2 @@
+import os, sys
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
