+++ gamedaybot/chat/discord.py
@@ -1,0 +1,36 @@
+import requests
+import json
+import logging
+
+logger = logging.getLogger(__name__)
+
+class DiscordException(Exception):
+    pass
+
+class Discord(object):
+    # Creates Discord Bot to send messages
+    def __init__(self, webhook_url):
+        self.webhook_url = webhook_url
+
+    def __repr__(self):
+        return "Discord Webhook Url(%s)" % self.webhook_url
+
+    def send_message(self, text):
+        # Sends a message to the chatroom
+        message = "```{0}```".format(text)
+        template = {
+            "content": message #limit 3000 chars
+        }
+
+        headers = {'content-type': 'application/json'}
+
+        if self.webhook_url not in (1, "1", ''):
+            r = requests.post(self.webhook_url,
+                              data=json.dumps(template), headers=headers)
+
+            if r.status_code != 204:
+                print(r.content)
+                logger.error(r.content)
+                raise DiscordException(r.content)
+
+            return r
